Provide a fallback when using sysfs fails
This commit "fixes" Lightbulb for people on custom kernels that do not
support the sysfs method that was added recently. This solution is not
optimal as the method it falls back to is the old, slow one but it is
better than just silently failing. This commit also adds some logging
to inform about these failures.

To kernel developers reading this: See the following change to the
hammerhead tree and add such a change to your init file as well, but
please review whether the path that permissions are granted for
matches the real path on the system before doing so.
https://github.com/AOSPA/android_device_lge_hammerhead/commit/26fde48a0964613eceb7ec73b8b7536c0223ce20

Change-Id: I172e60acb92aea02bf1b4ca88e1e02658570bf09